fix uniform test
