+++ src/harness/cu_pass/dpa_calculator/cbsds_creator/cbsd_height_distributor/cbsd_height_distributor.py
@@ -7,7 +7,8 @@ from cu_pass.dpa_calculator.cbsds_creator.cbsd_height_distributor.cbsd_height_ge
     LocationWithHeight
 from cu_pass.dpa_calculator.constants import REGION_TYPE_DENSE_URBAN, REGION_TYPE_RURAL, REGION_TYPE_SUBURBAN, \
     REGION_TYPE_URBAN
-from cu_pass.dpa_calculator.helpers.list_distributor import FractionalDistribution
+from cu_pass.dpa_calculator.helpers.list_distributor.fractional_distribution.fractional_distribution import \
+    FractionalDistribution
 from cu_pass.dpa_calculator.point_distributor import CoordinatesWithBearing
 from cu_pass.dpa_calculator.utilities import Point
